requirements.txt
Django==1.7.7
Pillow==2.8.1
argparse==1.2.1
beautifulsoup4==4.3.2
click==4.0
dj-database-url==0.3.0
dj-static==0.0.6
django-annoying==0.8.1
django-appconf==1.0.1
django-image-cropping==1.0.1
django-m2m-history==0.3.1
django-oauth-tokens==0.5.0
django-picklefield==0.3.1
django-rq==0.8.0
django-taggit==0.12.3
django-toolbelt==0.0.1
django-twitter-api==0.2.5
easy-thumbnails==2.2
gunicorn==19.3.0
oauthlib==0.7.2
psycopg2==2.6
redis==2.10.3
requests==2.6.0
requests-oauthlib==0.4.2
rq==0.5.4
six==1.9.0
static3==0.5.1
tweepy==3.3.0
wsgiref==0.1.2
